# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/genres": {
            "post": {
                "tags": [
                    "genres"
                ],
                "summary": "Store new genre",
                "description": "Returns genre data",
                "operationId": "storeGenre",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreGenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/genres/{id}": {
            "get": {
                "tags": [
                    "genres"
                ],
                "summary": "Get genre information",
                "description": "Returns genre data",
                "operationId": "getGenreById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "genres"
                ],
                "summary": "Update existing genre",
                "description": "Returns updated genre data",
                "operationId": "updateGenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGenreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "genres"
                ],
                "summary": "Delete existing genre",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteGenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Genre": {
                "title": "Genre",
                "description": "Genre model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "GenreResource": {
                "title": "GenreResource",
                "description": "Genre resource",
                "properties": {
                    "genre": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Genre"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GenreResource"
                }
            },
...
            "StoreGenreRequest": {
                "title": "Store Genre request",
                "description": "Store Genre request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new genre",
                        "type": "string",
                        "example": "A nice genre"
                    },
                    "": {
                        "title": "description",
                        "description": "Description of the new genre",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            },
...
            "UpdateGenreRequest": {
                "title": "Update Genre request",
                "description": "Update Genre request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new genre",
                        "type": "string",
                        "example": "A nice genre"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the new genre",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}